+++ R/get_predictions_glm.R
@@ -11,16 +11,25 @@ get_predictions_glm <- function(model, fitfram, ci.lvl, linv, value_adjustment, 
     # simulate predictions
     .do_simulate(model, terms, ci, ...)
   } else {
-    # for models from "robust"-pkg (glmRob) we need to
-    # suppress warnings about fake models
-    prdat <-
-      suppressWarnings(stats::predict.glm(
+    if (inherits(model, "bayesglm")) {
+      prdat <- stats::predict(
+        model,
+        newdata = fitfram,
+        type = "link",
+        se.fit = se,
+        ...
+      )
+    } else {
+      # for models from "robust"-pkg (glmRob) we need to
+      # suppress warnings about fake models
+      prdat <- suppressWarnings(stats::predict.glm(
         model,
         newdata = fitfram,
         type = "link",
         se.fit = se,
         ...
       ))
+    }
 
     # copy predictions
     .generic_prediction_data(model, fitfram, linv, prdat, se, ci.lvl, model_class, value_adjustment, terms, vcov.fun, vcov.type, vcov.args, condition, interval)